(feat) Add an entry for an enquiry for a non-zero companies house number
This is not a valid Atom feed, but doing this step-by-step, and I think the
consequences of an invalid feed are minor at this point.

'company_house_number' is allowed to be null in the database, and test data has
been found where it is the empty string. For now, just outputting a feed that
includes neither of these cases.