(feat) Make feed paginated
The pagination test is quite high level: the pagination must contain the
correct domain for the next page, and so a test of just the controller wasn't
deemed enough. It also ensures that URLs are properly encoded/decoded, to allow
for any ASCII character to be in the shared secret.