Activity stream stub feed (#254)
* (feat) Endpoint with empty JSON object

No security is needed at this point, since not even dummy secret information is
revealed.

* (feat) Add Activity Streams 2.0 Content-Type header

* (feat) Enforce that Authentication header must be sent with requests

At this point there is still no actual authentication done

* (feat) Hawk authentication, without nonce

The content_type and payload _are_ hashed, even if empty. Although

  - they're optional by the (rough) spec of Hawk;
  - the Ruby implementation used here tent/hawk-ruby allows it to be optional;
  - the python implementation kumar303/mohawk allows it to be optional;

we have found that

  - the Django REST framework implementation, kumar303/hawkrest, does not allow
    it to be optional.

To keep it simple, flexible, and have exactly one authentication mechanism
between languages, we make sure that the payload and content-type are hashed,
even if they're empty strings.

* (feat) Check nonce in Hawk authentication to mitigate risk of replay

* (feat) Dummy content for pen test to try to expose

* (feat) IP whitelisting on Activity Stream endpoint

The activity stream will be connecting from a short list of IP addresses. This
is to be able to add a small amount of extra protection over just Hawk.

* (refactor) Reduce duplication

* (refactor) Move nonce check out of index to reduce line length + cog complexity

* (refactor) Move code out of index

This is to move towards the CodeClimate max of 25 lines per method

* (refactor) Move code out of index

This is to move towards the CodeClimate max of 25 lines per method