(feat) Reduce Figaro usage to reduce memory usage for Activity Stream controller
Figaro seems to be allocating a _lot_ of objects, >200k after a few minutes.
Suspect the issue is similar to that in
https://github.com/laserlemon/figaro/issues/189

This change limits the scope of removing Figaro, in favour of just accessing
ENV directly, to just those which are done per-request on the Activiy Stream.
However, by changing application_controller, this would affect _all_ incoming
requests. Of course, hopefully for the better.

Note: Figaro loads configuration into ENV, which I think has no evidence of
bloat. It's just accessing the environment using the "conveniance" wrapper of
Figaro.env which is bloat-y. Also, a lookup for a key is O(N) on the number of
environment variables, which seems a bit unnecessary.