(fix) Incorrect pagination on opportunities endpoint
A where on the tuple (response_due_on, updated_at, id) > (a, b, c) would
return all rows where response_due_on > a, even if updated_at <= b. This
is not the correct behaviour for the pagination.

Note that without the index change, a full table scan occurred:

    EXPLAIN
      SELECT  "opportunities".* FROM "opportunities" WHERE (status=2
      AND response_due_on > to_timestamp(0.0)
      AND (updated_at, id) > (to_timestamp(0.0)

would include:

   Seq Scan on opportunities